requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
open_clip_torch>=2.20.0
Pillow>=9.5.0
